engine: KunMoe inference core - operators, kernels, attention, quantization, forward, generate, tokenizer, server
C99 inference engine for the KunMoe Mixture-of-Experts model. Targets
Pentium-class Win9x/DOS (gcc / Open Watcom, x87+MMX+SSE/SSE2) and
ARM926EJ-S (soft-float, integer operator tiers).

- LZ_FMT_T2 ternary-weight format and its scalar reference (base at
  dc0f08b); quantized int8/int16 operator tiers with an integer default
- four-path operator dispatch (float / SSE2 / MMX / ARM asm), q8/q4
  kernels, norm, GDN epilogue, conv1d, Hadamard, rotary, KV, attention
  (incl. speculative/MTP prefill), MoE routing, SSM branch
- fixed-point arithmetic where the target has no FPU; RTZ+FTZ is the
  default float semantics and x86 matches ARM byte-for-byte
- forward pass with the decimal place on 32-bit targets (lz_i64,
  LZ_PRI64); tokenizer/unicode/GBK; lz4d streaming model load;
  CLI (-i), HTTP/OpenAI server, shared LZSession front-end
- mathf, bf16, softfp, fwht, lz_int portability header (stdint floor)

Soft-float, BF16 and the int operator tiers are the ARM/no-FPU story;
see README.md for the target matrix.

+++ common/session.h
@@ -76,7 +76,7 @@ typedef struct {
     /* Last turn's statistics, written by lz_session_job, read after.
      * NOT reset between turns; a new job overwrites them. */
     int n_out, n_prompt_tok, reused;
-    double ms;
+    float ms;
 } LZSession;
 
 void lz_session_init(LZSession *s, LZModel *model, LZTokenizer *tok,
@@ -206,6 +206,6 @@ int lz_session_token_count(LZSession *s, const char *system,
 int   lz_session_last_reused(const LZSession *s);
 int   lz_session_last_n_out(const LZSession *s);
 int   lz_session_last_prompt_tok(const LZSession *s);
-double lz_session_last_ms(const LZSession *s);
+float lz_session_last_ms(const LZSession *s);
 
 #endif
